requirements.txt
fastapi==0.112.0
uvicorn[standard]==0.30.5
python-dotenv==1.0.0
httpx==0.24.1
requests==2.31.0
pytest==7.4.0
pytest-asyncio==0.21.0
pymysql==1.1.1
mysql-connector-python==9.0.0
openpyxl==3.1.5
tqdm==4.66.5
aiomysql==0.2.0
pandas==2.2.2
xlsxwriter==3.2.0
openai==1.51.2
pydantic